requirements.txt
flask==3.0.2
python-docx>=1.1.2,<2.0.0
openpyxl>=3.1.5,<4.0.0
python-pptx>=1.0.2
PyPDF2==3.0.1
markdown==3.5
Werkzeug>=3.0.0
python-magic==0.4.27
pywin32>=307; sys_platform == 'win32'
pytesseract==0.3.10
Pillow>=8.0.0
pdf2image==1.16.3
markitdown>=0.1.0
#-if you don't use docling, comment it out
docling
docling_core
docling-parse
docling-ibm-models
beautifulsoup4
filetype>=1.2.0
lxml>=4.9.3
marko
pydantic>=2.0.0
pydantic-settings>=2.0.0
pypdfium2
pylatexenc
rtree
scipy
tqdm>=4.65.0
typer
pluggy>=1.2.0
transformers>=4.42.0,<5.0.0
huggingface_hub>=0.23,<1
jsonlines>=3.1.0
numpy>=1.24.4
safetensors[torch]>=0.4.3
easyocr>=1.7
#- High-precision document converter
# marker-pdf
#magic-pdf[full]
python-magic-bin==0.4.14; sys_platform == 'win32'
requests>=2.32.3
pydub
accelerate
gunicorn>=20.1.0
markupsafe>=2.1.1 
typing_extensions>=4.5.0
flasgger==0.9.7.1
PyMuPDF>=1.19.6
jinja2==3.1.2
rich==12.6.0
html2text>=2020.1.16